Enemy
I have created a base enemy, they are fully animated with a randomized heath (Within a limit). They also have a variety of different sprites which will be randomly selected when they are spawned (Different colours).

I have also made sure that the projectile changes it's bitmap when it hits a collideable object.

I have created a MarkedForDeletion system into the game so at the end of the frame it will check to see what needs to be removed from the game to ensure the current frame functions correctly.

Next I will need to create a system which will allow for the enemies to randomly shoot back at the player.