Shooting Enemies
I have added a system which allows for the enemies to shoot and the player to be hit. I have also created a set path for the enemy to move alon; (left and right on the screen).

Next I will create a system which will randomly spawn an enemy and it will move onto the screen along a random path.