Enemies
I have made an array of enemies which has all of the functonality of a single enemy but allows for more than one enemy to be in the game at a single time. I have also created a lives and score UI which can dynamically change.

Next I aim to add in a pathfinding system which will spawn new enemies into the game allowing for the game to continue indefinatly.